Topo_Assessments/pyproject.toml
[project]
name = "topo-assessments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "camelot-py>=1.0.9",
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "fastmcp>=2.11.3",
    "ipython>=9.4.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "psycopg>=3.2.9",
    "psycopg2>=2.9.10",
    "python-pptx>=1.0.2",
]
